refactor: extract appointments request handling logic to controller
Since appointments controller needs patients and doctors data, wich is
not a duty of their controllers to return such data, added repository
modules for patients and doctors and make use of this repositories in
all of the controllers.

The appointments controller requires data about patients and doctors,
but it is not the responsibility of the patient and doctor controllers
to return such data. Therefore, created repository modules for patients
and doctors to return such data.

Now making use of these repositories in all of the relevant controllers.
For that, repository logic earlier found in controllers moved back to
repositories.